restructure response file
